backend/services/memory.py
import vertexai
from google.cloud import aiplatform
import os
from dotenv import load_dotenv

load_dotenv()

PROJECT_ID = os.getenv("GOOGLE_CLOUD_PROJECT")
LOCATION = os.getenv("GOOGLE_CLOUD_LOCATION", "us-central1")
AGENT_ENGINE_ID = os.getenv("AGENT_ENGINE_ID") # Set after deployment

def init_memory_bank():
    vertexai.init(project=PROJECT_ID, location=LOCATION)

async def store_insight(user_id: str, insight: str, source_url: str):
    """
    Store a new insight into the user's Memory Bank.
    """
    # This uses the generate_memories API pattern
    # In a real environment, you'd use the vertexai client
    print(f"Storing insight for {user_id}: {insight[:50]}...")
    return {"status": "stored", "user_id": user_id}

async def retrieve_relevant_context(user_id: str, query: str):
    """
    Retrieve relevant memories based on the search query.
    """
    # This uses the retrieve_memories API pattern
    print(f"Retrieving context for {user_id} related to: {query}")
    return [] # Placeholder for retrieved memories
